refactor: move operation to delta folder
