+++ shared-lib/lib-ot/src/core/delta/mod.rs
@@ -4,6 +4,7 @@ mod cursor;
 mod delta;
 mod delta_serde;
 mod iterator;
+pub mod operation;
 
 pub use builder::*;
 pub use cursor::*;
