+++ shared-lib/lib-ot/src/core/delta/operation/operation_serde.rs
@@ -1,4 +1,4 @@
-use crate::core::operation::{Attributes, Insert, Operation, Retain};
+use crate::core::delta::operation::{Attributes, Insert, Operation, Retain};
 use crate::core::ot_str::OTString;
 use serde::{
     de,
