+++ shared-lib/lib-ot/src/core/mod.rs
@@ -1,11 +1,10 @@
 mod delta;
 mod document;
 mod interval;
-mod operation;
 mod ot_str;
 
+pub use delta::operation::*;
 pub use delta::*;
 pub use document::*;
 pub use interval::*;
-pub use operation::*;
 pub use ot_str::*;
